[ci] Deflaky test_basic_5_client_mode (#40989)
---------
