+++ ci/ray_ci/core.tests.yml
@@ -13,7 +13,6 @@ flaky_tests:
   - //python/ray/tests:test_unhandled_error
   - //python/ray/tests:test_state_api_log
   - //python/ray/tests:test_ray_init_2
-  - //python/ray/tests:test_basic_5_client_mode
   - //python/ray/tests:test_client_builder
   - //python/ray/tests:test_client_library_integration
   - //python/ray/tests:test_object_assign_owner_client_mode
