+++ python/ray/tests/test_advanced_9.py
@@ -1,9 +1,11 @@
+import os
 import sys
 import time
 
 import pytest
 
 import ray
+import ray._private.ray_constants as ray_constants
 from ray._private.test_utils import (
     Semaphore,
     enable_external_redis,
@@ -207,8 +209,6 @@ def test_node_liveness_after_restart(ray_start_cluster):
 def test_worker_oom_score(shutdown_only):
     @ray.remote
     def get_oom_score():
-        import os
-
         pid = os.getpid()
         with open(f"/proc/{pid}/oom_score", "r") as f:
             oom_score = f.read()
@@ -382,7 +382,6 @@ def test_redis_wrong_password(monkeypatch, external_redis, call_ray_stop_only):
 
 @pytest.mark.skipif(not enable_external_redis(), reason="Only valid in redis env")
 def test_redis_full(ray_start_cluster_head):
-    import os
     import redis
 
     gcs_address = ray_start_cluster_head.gcs_address
@@ -449,8 +448,6 @@ def test_gcs_fd_usage(shutdown_only):
     @ray.remote(runtime_env={"env_vars": {"Hello": "World"}})
     class A:
         def f(self):
-            import os
-
             return os.environ.get("Hello")
 
     # In case there are still some pre-start workers, consume all of them
@@ -473,9 +470,50 @@ def test_gcs_fd_usage(shutdown_only):
     assert (new_fd_num - base_fd_num) <= len(bb) * 2 + 1
 
 
+@pytest.mark.skipif(
+    sys.platform != "linux", reason="jemalloc is only prebuilt on linux"
+)
+def test_jemalloc_ray_start(monkeypatch, ray_start_cluster):
+    def check_jemalloc_enabled(pid=None):
+        if pid is None:
+            pid = os.getpid()
+        pmap = subprocess.run(
+            ["pmap", str(pid)], check=True, text=True, stdout=subprocess.PIPE
+        )
+        return "libjemalloc.so" in pmap.stdout
+
+    # Firstly, remove the LD_PRELOAD and make sure
+    # jemalloc is loaded.
+    monkeypatch.delenv("LD_PRELOAD", False)
+    cluster = ray_start_cluster
+    node = cluster.add_node(num_cpus=1)
+
+    # Make sure raylet/gcs/worker all have jemalloc
+    assert check_jemalloc_enabled(
+        node.all_processes[ray_constants.PROCESS_TYPE_GCS_SERVER][0].process.pid
+    )
+    assert check_jemalloc_enabled(
+        node.all_processes[ray_constants.PROCESS_TYPE_RAYLET][0].process.pid
+    )
+    assert not ray.get(ray.remote(check_jemalloc_enabled).remote())
+
+    ray.shutdown()
+    cluster.shutdown()
+
+    monkeypatch.setenv("LD_PRELOAD", "")
+    node = cluster.add_node(num_cpus=1)
+    # Make sure raylet/gcs/worker all have jemalloc
+    assert not check_jemalloc_enabled(
+        node.all_processes[ray_constants.PROCESS_TYPE_GCS_SERVER][0].process.pid
+    )
+    assert not check_jemalloc_enabled(
+        node.all_processes[ray_constants.PROCESS_TYPE_RAYLET][0].process.pid
+    )
+    assert not ray.get(ray.remote(check_jemalloc_enabled).remote())
+
+
 if __name__ == "__main__":
     import pytest
-    import os
 
     if os.environ.get("PARALLEL_CI"):
         sys.exit(pytest.main(["-n", "auto", "--boxed", "-vs", __file__]))
